+++ src/server/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod auth;
 pub(crate) mod deployment_summary;
 pub(crate) mod grpc_server;
 pub(crate) mod web_api_server;